Commit Todey 26/1/2018 -ScrollingActivity
