+++ app/src/main/AndroidManifest.xml
@@ -9,18 +9,16 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="com.example.potikorn.testcoopapp.MainActivity"></activity>
-
+        <activity android:name=".MainActivity" />
         <activity
-            android:name=".MainActivityFragmentation"
-            android:label="@string/title_activity_main_fragmentation">
+            android:name=".ScrollingActivity"
+            android:label="@string/title_activity_scrolling"
+            android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-
     </application>
 
 </manifest>
