+++ app/src/main/res/menu/navigation.xml
@@ -2,18 +2,13 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
+        android:titleCondensed="#fffafa"
         android:id="@+id/navigation_home"
-        android:icon="@drawable/ic_home_black_24dp"
         android:title="@string/title_home" />
 
     <item
+        android:titleCondensed="#fffafa"
         android:id="@+id/navigation_dashboard"
-        android:icon="@drawable/ic_dashboard_black_24dp"
         android:title="@string/title_dashboard" />
-
-    <item
-        android:id="@+id/navigation_notifications"
-        android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/title_notifications" />
 
 </menu>
