+++ app/src/main/res/values/colors.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#707070</color>
-    <color name="colorPrimaryDark">#707070</color>
+    <color name="colorPrimary">#2e2d2d</color>
+    <color name="colorPrimaryDark">#262525</color>
     <color name="colorAccent">#fffa6b</color>
 
     <color name="colorBlack">#555555</color>
